refactor(controllers): update imports to use refactored widgets
- Update settings_controller.dart to import from messages/messages.dart
- Update url_launch_controller.dart to import from messages/messages.dart
- Maintain functionality while using new widget structure
