+++ lib/app/core/controllers/settings_controller.dart
@@ -1,7 +1,7 @@
 import 'dart:io';
 import 'package:flutter/foundation.dart';
 import '../logger/app_logger.dart';
-import '../ui/widgets/messages.dart';
+import '../ui/widgets/messages/messages.dart';
 
 class SettingsController extends ChangeNotifier {
   SettingsController({
@@ -21,25 +21,6 @@ class SettingsController extends ChangeNotifier {
     } catch (e) {
       _logger.error('Erro ao encerrar o aplicativo: $e');
       Messages.alert('Erro ao encerrar o aplicativo');
-    }
-  }
-
-  // Método para mutar o áudio do WebView
-  Future<void> muteAppAudio() async {
-    try {
-      if (Platform.isWindows) {
-        Messages.info('Funcionalidade ainda não implementada');
-      }
-      // else if (Platform.isMacOS) {
-      //   await shell.run('osascript -e "set volume output muted true"');
-      // } else if (Platform.isLinux) {
-      //   await shell.run('amixer -q -D pulse sset Master mute');
-      // } else {
-      //   throw 'Plataforma não suportada para mutar áudio';
-      // }
-    } catch (e) {
-      _logger.error('Erro ao mutar o áudio do aplicativo: $e');
-      Messages.alert('Erro ao mutar o áudio do aplicativo');
     }
   }
 }
